Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,8 +1,8 @@
 #include<stdio.h>
 int main()
 {
-    int a,b,c;
-    scanf("%d\t%d\t%d",&a,&b,&c);
+    float a,b,c;
+    scanf("%f\t%f\t%f",&a,&b,&c);
     float e;
     e=(a+b+c)/3;
     printf("Average: %.2f",e);
